fix: checksum error

+++ protocols.cpp
@@ -141,7 +141,7 @@ bool verify_icmp_checksum(icmp *msg, int buf_len)
         total += (msg->data[buf_len - 1] << 8);
     }
     finish_checksum(&total, reinterpret_cast<uint16_t *>(&msg->type), buf_len - sizeof(ip_header));
-    return total == 0;
+    return static_cast<uint16_t>(total) == 0;
 }
 
 bool verify_ip_checksum(ip_header *msg)
@@ -149,5 +149,5 @@ bool verify_ip_checksum(ip_header *msg)
     unsigned int total = 0;
     uint16_t *header = reinterpret_cast<uint16_t *>(&msg->version_ihl);
     finish_checksum(&total, header, sizeof(ip_header) - sizeof(ethernet_header));
-    return total == 0;
+    return static_cast<uint16_t>(total) == 0;
 }
